Removed ExprVar
Using ScalarExprVar and Vec3ExprVar and Mat33ExprVar again.

Having the generic one made it hard to keep track of the expression types.
It was easy enough to add ExprTypeHelpers that were needed so that we
could still use these *ExprVar types in expressions without having
to explicitly use expr().